@@ -128,89 +111,91 @@
     "publier.loire-atlantique": "tsx src/index.ts publier -z \"fr:departement:44\" -m \"./assets/output/loire-atlantique/publier.json\"",
     "publier.rhin-occ": "tsx src/index.ts publier -z \"fr:region:76\" -m \"./assets/output/rhin-occ/publier.json\"",
     "publier.res-in": "tsx src/index.ts publier -z \"fr:commune:69123\" -m \"./assets/output/res-in/publier.json\"",
-    "publier.sarthe": "tsx src/index.ts publier -z \"fr:departement:72\" -m \"./assets/output/sarthe/publier.json\"",
     "publier.siilab": "tsx src/index.ts publier -z \"fr:region:32\" -m \"./assets/output/siilab/publier.json\"",
     "publier.ultra-num": "tsx src/index.ts publier -z \"fr:region:04\" -m \"./assets/output/ultra-num/publier.json\"",
-    "dedupliquer.all": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false\" -n \"Cartographie nationale des lieux d'inclusion numérique\" -t National -o ./assets/output/cartographie-nationale",
-    "dedupliquer.aidants-connect": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Aidants Connect\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Aidants Connect\" -n \"Aidants Connect\" -t \"National\" -o \"./assets/output/aidants-connect\" -i true",
-    "dedupliquer.aix-en-provence": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Mairie-Aix-en-Provence\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Mairie-Aix-en-Provence\" -n \"Mairie Aix en Provence\" -t \"Bouches-du-Rhône\" -o \"./assets/output/aix-en-provence\" -i true",
-    "dedupliquer.angers": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Angers\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Angers\" -n \"Angers\" -t \"Maine-et-Loire\" -o \"./assets/output/angers\" -i true",
-    "dedupliquer.bus-france-services-charente": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Bus France Services en Charente\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Bus France Services en Charente\" -n \"Bus France Services en Charente\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/bus-france-services-charente\" -i true",
-    "dedupliquer.charente-maritime": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Département de la Charente-Maritime\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Département de la Charente-Maritime\" -n \"Département de la Charente-Maritime\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/charente-maritime\" -i true",
-    "dedupliquer.conseiller-numerique": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Conseiller Numerique\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Conseiller Numerique\" -n \"Conseiller Numerique\" -t \"National\" -o \"./assets/output/conseiller-numerique\" -i true",
-    "dedupliquer.coop-numerique": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Coop numérique\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Coop numérique\" -n \"Coop numérique\" -t \"National\" -o \"./assets/output/coop-numerique\" -i true",
-    "dedupliquer.corse": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Corse\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Corse\" -n \"Corse\" -t \"Corse\" -o \"./assets/output/corse\" -i true",
-    "dedupliquer.dora": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=dora\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=dora\" -n \"dora\" -t \"National\" -o \"./assets/output/dora\" -i true",
-    "dedupliquer.epernay": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Epernay\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Epernay\" -n \"Epernay\" -t \"Marne\" -o \"./assets/output/epernay\" -i true",
-    "dedupliquer.etapes-numerique": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Etapes Numerique\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Etapes Numerique\" -n \"Etapes Numerique\" -t \"National\" -o \"./assets/output/etapes-numerique\" -i true",
-    "dedupliquer.fibre-64": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Fibre 64\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Fibre 64\" -n \"Fibre 64\" -t \"Pyrenees-Atlantique\" -o \"./assets/output/fibre-64\" -i true",
-    "dedupliquer.france-services": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=France Services\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=France Services\" -n \"France Services\" -t \"National\" -o \"./assets/output/france-services\" -i true",
-    "dedupliquer.francil-in": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Francil-in\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Francil-in\" -n \"Francil-in\" -t \"Île-de-France\" -o \"./assets/output/francil-in\" -i true",
-    "dedupliquer.france-tiers-lieux": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=France tiers-lieux\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=France tiers-lieux\" -n \"France tiers-lieux\" -t \"National\" -o \"./assets/output/france-tiers-lieux\" -i true",
-    "dedupliquer.gironde": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Gironde\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Gironde\" -n \"Gironde\" -t\"Nouvelle-Aquitaine\" -o \"./assets/output/gironde\" -i true",
-    "dedupliquer.grand-paris-sud": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Grand Paris Sud\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Grand Paris Sud\" -n \"Grand Paris Sud\" -t\"Île-de-France\" -o \"./assets/output/grand-paris-sud\" -i true",
-    "dedupliquer.haute-vienne": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Haute-Vienne\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Haute-Vienne\" -n \"Haute-Vienne\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/haute-vienne\" -i true",
-    "dedupliquer.hinaura": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Hinaura\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Hinaura\" -n \"Hinaura\" -t \"Auvergne-Rhône-Alpes\" -o \"./assets/output/hinaura\" -i true",
-    "dedupliquer.hub-bretagne": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Hub Bretagne\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Hub Bretagne\" -n \"Hub Bretagne\" -t \"Bretagne\" -o \"./assets/output/hub-bretagne\" -i true",
-    "dedupliquer.hub-lo": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Hub-lo\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Hub-lo\" -n \"Hub-lo\" -t \"Centre-Val-de-Loire\" -o \"./assets/output/hub-lo\" -i true",
-    "dedupliquer.maine-et-loire": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Département du Maine-et-Loire\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Département du Maine-et-Loire\" -n \"Département du Maine-et-Loire\" -t \"Maine-et-Loire\" -o \"./assets/output/maine-et-loire\" -i true",
-    "dedupliquer.mednum-bfc": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Mednum BFC\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Mednum BFC\" -n \"Mednum BFC\" -t \"Bourgogne-Franche-Comté\" -o \"./assets/output/mednum-bfc\" -i true",
-    "dedupliquer.mulhouse": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Mulhouse\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Mulhouse\" -n \"Mulhouse\" -t \"Haut-Rhin\" -o \"./assets/output/mulhouse\" -i true",
-    "dedupliquer.nouvelle-caledonie": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Nouvelle-Calédonie\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Nouvelle-Calédonie\" -n \"Nouvelle-Calédonie\" -t \"Nouvelle-Calédonie\" -o \"./assets/output/nouvelle-caledonie\" -i true",
-    "dedupliquer.vendee": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Vendée\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Vendée\" -n \"Vendée\" -t \"Vendée\" -o \"./assets/output/vendee\" -i true",
-    "dedupliquer.la-creuse": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=La Creuse\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=La Creuse\" -n \"La Creuse\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/la-creuse\" -i true",
-    "dedupliquer.le-havre": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Le Havre\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Le Havre\" -n \"Le Havre\" -t \"Le Havre\" -o \"./assets/output/le-havre\" -i true",
-    "dedupliquer.les-landes": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Les Landes\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Les Landes\" -n \"Les Landes\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/les-landes\" -i true",
-    "dedupliquer.paca": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Paca\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Paca\" -n \"Paca\" -t \"Provence-Alpes-Côte d'Azur\" -o \"./assets/output/paca\" -i true",
-    "dedupliquer.loire-atlantique": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Loire Atlantique\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Loire Atlantique\" -n \"Loire Atlantique\" -t \"Loire-Atlantique\" -o \"./assets/output/loire-atlantique\" -i true",
-    "dedupliquer.rhin-occ": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=RhinOcc\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=RhinOcc\" -n \"RhinOcc\" -t \"Occitanie\" -o \"./assets/output/rhin-occ\" -i true",
-    "dedupliquer.res-in": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Res-in\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Res-in\" -n \"Res-in\" -t \"Lyon\" -o \"./assets/output/res-in\" -i true",
-    "dedupliquer.sarthe": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=La Sarthe\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=La Sarthe\" -n \"La Sarthe\" -t \"Pays de la Loire\" -o \"./assets/output/sarthe\" -i true",
-    "dedupliquer.siilab": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Siilab\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Siilab\" -n \"Siilab\" -t \"Hauts-de-France\" -o \"./assets/output/siilab\" -i true",
-    "dedupliquer.ultra-num": "tsx src/index.ts dedupliquer -s \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?deduplicated[exists]=false&source[eq]=Ultra-numerique\" -b \"https://cartographie.societenumerique.gouv.fr/api/v0/lieux-inclusion-numerique/with-duplicates?mergedIds[exists]=false&source[eq]=Ultra-numerique\" -n \"Ultra-numerique\" -t \"La Reunion\" -o \"./assets/output/ultra-num\" -i true",
+    "dedupliquer.aidants-connect": "tsx src/index.ts dedupliquer -s \"./assets/output/aidants-connect/*-aidants-connect-lieux-de-mediation-numeriques-national.json\" -b \"./assets/output/aidants-connect/*-aidants-connect-lieux-de-mediation-numeriques-national.json\" -n \"Aidants Connect\" -t \"National\" -o \"./assets/output/aidants-connect\" -i true",
+    "dedupliquer.aix-en-provence": "tsx src/index.ts dedupliquer -s \"./assets/output/aix-en-provence/*-mairie-aix-en-provence-lieux-de-mediation-numeriques-bouches-du-rhone.json\" -b \"./assets/output/aix-en-provence/*-mairie-aix-en-provence-lieux-de-mediation-numeriques-bouches-du-rhone.json\" -n \"Mairie Aix en Provence\" -t \"Bouches-du-Rhône\" -o \"./assets/output/aix-en-provence\" -i true",
+    "dedupliquer.angers": "tsx src/index.ts dedupliquer -s \"./assets/output/angers/*-angers-lieux-de-mediation-numeriques-maine-et-loire.csv\" -b \"./assets/output/angers/*-angers-lieux-de-mediation-numeriques-maine-et-loire.csv\" -n \"Angers\" -t \"Maine-et-Loire\" -o \"./assets/output/angers\" -i true",
+    "dedupliquer.bus-france-services-charente": "tsx src/index.ts dedupliquer -s \"./assets/output/bus-france-services-charente/*-bus-france-services-en-charente-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -b \"./assets/output/bus-france-services-charente/*-bus-france-services-en-charente-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -n \"Bus France Services en Charente\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/bus-france-services-charente\" -i true",
+    "dedupliquer.charente-maritime": "tsx src/index.ts dedupliquer -s \"./assets/output/charente-maritime/*-departement-de-la-charente-maritime-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -b \"./assets/output/charente-maritime/*-departement-de-la-charente-maritime-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -n \"Département de la Charente-Maritime\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/charente-maritime\" -i true",
+    "dedupliquer.conseiller-numerique": "tsx src/index.ts dedupliquer -s \"./assets/output/conseiller-numerique/*-conseiller-numerique-lieux-de-mediation-numeriques-national.json\" -b \"./assets/output/conseiller-numerique/*-conseiller-numerique-lieux-de-mediation-numeriques-national.json\" -n \"Conseiller Numerique\" -t \"National\" -o \"./assets/output/conseiller-numerique\" -i true",
+    "dedupliquer.coop-numerique": "tsx src/index.ts dedupliquer -s \"./assets/output/coop-numerique/*-coop-numerique-lieux-de-mediation-numeriques-national.json\" -b \"./assets/output/coop-numerique/*-coop-numerique-lieux-de-mediation-numeriques-national.json\" -n \"Coop numérique\" -t \"National\" -o \"./assets/output/coop-numerique\" -i true",
+    "dedupliquer.corse": "tsx src/index.ts dedupliquer -s \"./assets/output/corse/*-corse-lieux-de-mediation-numeriques-corse.json\" -b \"./assets/output/corse/*-corse-lieux-de-mediation-numeriques-corse.json\" -n \"Corse\" -t \"Corse\" -o \"./assets/output/corse\" -i true",
+    "dedupliquer.dora": "tsx src/index.ts dedupliquer -s \"./assets/output/dora/*-dora-lieux-de-mediation-numeriques-national.json\" -b \"./assets/output/dora/*-dora-lieux-de-mediation-numeriques-national.json\" -n \"dora\" -t \"National\" -o \"./assets/output/dora\" -i true",
+    "dedupliquer.epernay": "tsx src/index.ts dedupliquer -s \"./assets/output/epernay/*-epernay-lieux-de-mediation-numeriques-marne.json\" -b \"./assets/output/epernay/*-epernay-lieux-de-mediation-numeriques-marne.json\" -n \"Epernay\" -t \"Marne\" -o \"./assets/output/epernay\" -i true",
+    "dedupliquer.etapes-numerique": "tsx src/index.ts dedupliquer -s \"./assets/output/etapes-numerique/*-etapes-numerique-lieux-de-mediation-numeriques-national.json\" -b \"./assets/output/etapes-numerique/*-etapes-numerique-lieux-de-mediation-numeriques-national.json\" -n \"Etapes Numerique\" -t \"National\" -o \"./assets/output/etapes-numerique\" -i true",
+    "dedupliquer.fibre-64": "tsx src/index.ts dedupliquer -s \"./assets/output/fibre-64/*-fibre-64-lieux-de-mediation-numeriques-pyrenees-atlantique.json\" -b \"./assets/output/fibre-64/*-fibre-64-lieux-de-mediation-numeriques-pyrenees-atlantique.json\" -n \"Fibre 64\" -t \"Pyrenees-Atlantique\" -o \"./assets/output/fibre-64\" -i true",
+    "dedupliquer.france-services": "tsx src/index.ts dedupliquer -s \"./assets/output/france-services/*-france-services-lieux-de-mediation-numeriques-national.json\" -b \"./assets/output/france-services/*-france-services-lieux-de-mediation-numeriques-national.json\" -n \"France Services\" -t \"National\" -o \"./assets/output/france-services\" -i true",
+    "dedupliquer.francil-in": "tsx src/index.ts dedupliquer -s \"./assets/output/francil-in/*-francil-in-lieux-de-mediation-numeriques-ile-de-france.json\" -b \"./assets/output/francil-in/*-francil-in-lieux-de-mediation-numeriques-ile-de-france.json\" -n \"Francil-in\" -t \"Île-de-France\" -o \"./assets/output/francil-in\" -i true",
+    "dedupliquer.gironde": "tsx src/index.ts dedupliquer -s \"./assets/output/gironde/*-gironde-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -b \"./assets/output/gironde/*-gironde-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -n \"Gironde\" -t\"Nouvelle-Aquitaine\" -o \"./assets/output/gironde\" -i true",
+    "dedupliquer.grand-paris-sud": "tsx src/index.ts dedupliquer -s \"./assets/output/grand-paris-sud/*-grand-paris-sud-lieux-de-mediation-numeriques-ile-de-france.json\" -b \"./assets/output/grand-paris-sud/*-grand-paris-sud-lieux-de-mediation-numeriques-ile-de-france.json\" -n \"Grand Paris Sud\" -t\"Île-de-France\" -o \"./assets/output/grand-paris-sud\" -i true",
+    "dedupliquer.haute-vienne": "tsx src/index.ts dedupliquer -s \"./assets/output/haute-vienne/*-haute-vienne-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -b \"./assets/output/haute-vienne/*-haute-vienne-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -n \"Haute-Vienne\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/haute-vienne\" -i true",
+    "dedupliquer.hinaura": "tsx src/index.ts dedupliquer -s \"./assets/output/hinaura/*-hinaura-lieux-de-mediation-numeriques-auvergne-rhone-alpes.json\" -b \"./assets/output/hinaura/*-hinaura-lieux-de-mediation-numeriques-auvergne-rhone-alpes.json\" -n \"Hinaura\" -t \"Auvergne-Rhône-Alpes\" -o \"./assets/output/hinaura\" -i true",
+    "dedupliquer.hub-bretagne": "tsx src/index.ts dedupliquer -s \"./assets/output/hub-bretagne/*-hub-bretagne-lieux-de-mediation-numeriques-bretagne.json\" -b \"./assets/output/hub-bretagne/*-hub-bretagne-lieux-de-mediation-numeriques-bretagne.json\" -n \"Hub Bretagne\" -t \"Bretagne\" -o \"./assets/output/hub-bretagne\" -i true",
+    "dedupliquer.hub-lo": "tsx src/index.ts dedupliquer -s \"./assets/output/hub-lo/*-hub-lo-lieux-de-mediation-numeriques-centre-val-de-loire.json\" -b \"./assets/output/hub-lo/*-hub-lo-lieux-de-mediation-numeriques-centre-val-de-loire.json\" -n \"Hub-lo\" -t \"Centre-Val-de-Loire\" -o \"./assets/output/hub-lo\" -i true",
+    "dedupliquer.la-creuse": "tsx src/index.ts dedupliquer -s \"./assets/output/la-creuse/*-la-creuse-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -b \"./assets/output/la-creuse/*-la-creuse-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -n \"La Creuse\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/la-creuse\" -i true",
+    "dedupliquer.le-havre": "tsx src/index.ts dedupliquer -s \"./assets/output/le-havre/*-le-havre-lieux-de-mediation-numeriques-le-havre.json\" -b \"./assets/output/le-havre/*-le-havre-lieux-de-mediation-numeriques-le-havre.json\" -n \"Le Havre\" -t \"Le Havre\" -o \"./assets/output/le-havre\" -i true",
+    "dedupliquer.les-landes": "tsx src/index.ts dedupliquer -s \"./assets/output/les-landes/*-les-landes-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -b \"./assets/output/les-landes/*-les-landes-lieux-de-mediation-numeriques-nouvelle-aquitaine.json\" -n \"Les Landes\" -t \"Nouvelle-Aquitaine\" -o \"./assets/output/les-landes\" -i true",
+    "dedupliquer.loire-atlantique": "tsx src/index.ts dedupliquer -s \"./assets/output/loire-atlantique/*-loire-atlantique-lieux-de-mediation-numeriques-loire-atlantique.json\" -b \"./assets/output/loire-atlantique/*-loire-atlantique-lieux-de-mediation-numeriques-loire-atlantique.json\" -n \"Loire Atlantique\" -t \"Loire-Atlantique\" -o \"./assets/output/loire-atlantique\" -i true",
+    "dedupliquer.maine-et-loire": "tsx src/index.ts dedupliquer -s \"./assets/output/maine-et-loire/*-departement-du-maine-et-loire-lieux-de-mediation-numeriques-maine-et-loire.json\" -b \"./assets/output/maine-et-loire/*-departement-du-maine-et-loire-lieux-de-mediation-numeriques-maine-et-loire.json\" -n \"Département du Maine-et-Loire\" -t \"Maine-et-Loire\" -o \"./assets/output/maine-et-loire\" -i true",
+    "dedupliquer.mednum-bfc": "tsx src/index.ts dedupliquer -s \"./assets/output/mednum-bfc/*-mednum-bfc-lieux-de-mediation-numeriques-bourgogne-franche-comte.json\" -b \"./assets/output/mednum-bfc/*-mednum-bfc-lieux-de-mediation-numeriques-bourgogne-franche-comte.json\" -n \"Mednum BFC\" -t \"Bourgogne-Franche-Comté\" -o \"./assets/output/mednum-bfc\" -i true",
+    "dedupliquer.mulhouse": "tsx src/index.ts dedupliquer -s \"./assets/output/mulhouse/*-mulhouse-lieux-de-mediation-numeriques-haut-rhin.json\" -b \"./assets/output/mulhouse/*-mulhouse-lieux-de-mediation-numeriques-haut-rhin.json\" -n \"Mulhouse\" -t \"Haut-Rhin\" -o \"./assets/output/mulhouse\" -i true",
+    "dedupliquer.nouvelle-caledonie": "tsx src/index.ts dedupliquer -s \"./assets/output/nouvelle-caledonie/*-nouvelle-caledonie-lieux-de-mediation-numeriques-nouvelle-caledonie.json\" -b \"./assets/output/nouvelle-caledonie/*-nouvelle-caledonie-lieux-de-mediation-numeriques-nouvelle-caledonie.json\" -n \"Nouvelle-Calédonie\" -t \"Nouvelle-Calédonie\" -o \"./assets/output/nouvelle-caledonie\" -i true",
+    "dedupliquer.paca": "tsx src/index.ts dedupliquer -s \"./assets/output/paca/*-paca-lieux-de-mediation-numeriques-provence-alpes-cote-d-azur.json\" -b \"./assets/output/paca/*-paca-lieux-de-mediation-numeriques-provence-alpes-cote-d-azur.json\" -n \"Paca\" -t \"Provence-Alpes-Côte d'Azur\" -o \"./assets/output/paca\" -i true",
+    "dedupliquer.rhin-occ": "tsx src/index.ts dedupliquer -s \"./assets/output/rhin-occ/*-rhinocc-lieux-de-mediation-numeriques-occitanie.json\" -b \"./assets/output/rhin-occ/*-rhinocc-lieux-de-mediation-numeriques-occitanie.json\" -n \"RhinOcc\" -t \"Occitanie\" -o \"./assets/output/rhin-occ\" -i true",
+    "dedupliquer.res-in": "tsx src/index.ts dedupliquer -s \"./assets/output/____/*-____.json\" -b \"./assets/output/____/*-____.json\" -n \"Res-in\" -t \"Lyon\" -o \"./assets/output/res-in\" -i true",
+    "dedupliquer.siilab": "tsx src/index.ts dedupliquer -s \"./assets/output/siilab/*-siilab-lieux-de-mediation-numeriques-hauts-de-france.json\" -b \"./assets/output/siilab/*-siilab-lieux-de-mediation-numeriques-hauts-de-france.json\" -n \"Siilab\" -t \"Hauts-de-France\" -o \"./assets/output/siilab\" -i true",
+    "dedupliquer.ultra-num": "tsx src/index.ts dedupliquer -s \"./assets/output/____/*-____.json\" -b \"./assets/output/____/*-____.json\" -n \"Ultra-numerique\" -t \"La Reunion\" -o \"./assets/output/ultra-num\" -i true",
+    "dedupliquer.vendee": "tsx src/index.ts dedupliquer -s \"./assets/output/vendee/*-vendee-lieux-de-mediation-numeriques-vendee.json\" -b \"./assets/output/vendee/*-vendee-lieux-de-mediation-numeriques-vendee.json\" -n \"Vendée\" -t \"Vendée\" -o \"./assets/output/vendee\" -i true",
+    "dedupliquer.merged-json": "tsx src/index.ts dedupliquer -s ./assets/to-deduplicate/merged_output.json -b ./assets/to-deduplicate/merged_output.json -n \"Données dédupliquées\" -t National -o ./assets/deduplicated",
     "dedupliquer.merged-csv": "tsx src/index.ts dedupliquer -s ./merged_output.csv -b ./merged_output.csv -n \"Données dédupliquées\" -t National -o ./assets/deduplicated",
-    "fusionner": "tsx src/index.ts fusionner -i \"./assets/output/**/*.csv\" -o \"./assets/merged\""
+    "fusionner": "tsx src/index.ts fusionner -i \"./to-merge/*-sans-doublons.json\" -o \"./merged\"",
+    "prepare": "husky"
   },
   "devDependencies": {
-    "@commitlint/cli": "^19.3.0",
-    "@commitlint/config-conventional": "^19.2.2",
+    "@commitlint/cli": "^19.6.1",
+    "@commitlint/config-conventional": "^19.6.0",
+    "@eslint/js": "^9.17.0",
     "@turf/helpers": "^7.1.0",
+    "@types/form-data": "^2.5.2",
+    "@types/flat": "^5.0.5",
+    "@types/geojson": "^1.0.6",
     "@types/glob": "^8.1.0",
-    "@types/inquirer": "^9.0.3",
-    "@types/jest": "^29.1.1",
-    "@types/node": "^20.14.6",
-    "@typescript-eslint/eslint-plugin": "^5.39.0",
-    "@typescript-eslint/parser": "^5.39.0",
-    "eslint": "^8.24.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-jest": "^27.1.0",
-    "husky": "^9.0.11",
+    "@types/inquirer": "^9.0.7",
+    "@types/node": "^22.10.2",
+    "@types/proj4": "^2.5.5",
+    "@typescript-eslint/eslint-plugin": "^8.18.1",
+    "@typescript-eslint/parser": "^8.18.1",
+    "eslint": "^9.17.0",
+    "eslint-config-prettier": "^9.1.0",
+    "globals": "^15.14.0",
+    "husky": "^9.1.7",
     "iconv-lite": "^0.6.3",
-    "jest": "^29.1.2",
-    "lint-staged": "^15.2.7",
+    "lint-staged": "^15.2.11",
     "prettier": "^3.3.2",
     "standard-version": "^9.5.0",
-    "ts-jest": "^29.0.3",
     "ts-node": "^10.9.2",
-    "tsx": "^4.19.1"
+    "tsx": "^4.19.1",
+    "typescript-eslint": "^8.18.1",
+    "vitest": "^2.1.8"
   },
   "dependencies": {
-    "@gouvfr-anct/lieux-de-mediation-numerique": "^2.0.1",
-    "@gouvfr-anct/timetable-to-osm-opening-hours": "^1.0.1",
-    "@turf/turf": "^6.5.0",
+    "@gouvfr-anct/lieux-de-mediation-numerique": "^2.1.0",
+    "@gouvfr-anct/timetable-to-osm-opening-hours": "^2.1.0",
+    "@turf/turf": "^7.1.0",
     "axios": "^1.7.7",
     "axios-retry": "^4.5.0",
-    "commander": "^10.0.0",
-    "crypto": "^1.0.1",
+    "commander": "^12.1.0",
     "csv-parse": "^5.5.6",
     "csv-stringify": "^6.5.1",
     "csvtojson": "^2.0.10",
     "dotenv": "^16.0.3",
-    "flat": "^5.0.2",
+    "flat": "^6.0.1",
     "fuzzball": "^2.1.2",
-    "glob": "^10.4.5",
-    "inquirer": "^8.0.0",
+    "glob": "^11.0.0",
+    "inquirer": "^9.3.7",
     "proj4": "^2.9.0",
+    "form-data": "^4.0.1",
     "typescript": "^5.4.5"
   }
 }
