@@ -1,5 +1,3 @@
-/* eslint-disable max-lines-per-function */
-
 import { SchemaLieuMediationNumerique } from '@gouvfr-anct/lieux-de-mediation-numerique';
 
 const HEADERS: (keyof SchemaLieuMediationNumerique)[] = [
@@ -42,32 +40,32 @@ const toDoubleQuoted = (header?: string): string => (header == null ? '' : `"${h
 const fieldsArrayFrom = (lieuMediationNumerique: SchemaLieuMediationNumerique): (string | undefined)[] => [
   lieuMediationNumerique.id,
   lieuMediationNumerique.pivot,
-  lieuMediationNumerique.nom.replace(/"/gu, '').replace(/\n/gu, ''),
+  lieuMediationNumerique.nom.replace(/"/g, '').replace(/\n/g, ''),
   lieuMediationNumerique.commune,
   lieuMediationNumerique.code_postal,
   lieuMediationNumerique.code_insee,
-  lieuMediationNumerique.adresse.replace(/"/gu, '').replace(/\n/gu, '').replace(/\s$/gu, '').replace(/\s+/gu, ' '),
-  lieuMediationNumerique.complement_adresse?.replace(/"/gu, '').replace(/\n/gu, '').replace(/\s$/gu, '').replace(/\s+/gu, ' '),
+  lieuMediationNumerique.adresse.replace(/"/g, '').replace(/\n/g, '').replace(/\s$/g, '').replace(/\s+/g, ' '),
+  lieuMediationNumerique.complement_adresse?.replace(/"/g, '').replace(/\n/g, '').replace(/\s$/g, '').replace(/\s+/g, ' '),
   lieuMediationNumerique.latitude?.toString(),
   lieuMediationNumerique.longitude?.toString(),
   lieuMediationNumerique.typologie,
   lieuMediationNumerique.telephone,
   lieuMediationNumerique.courriels,
   lieuMediationNumerique.site_web,
-  lieuMediationNumerique.horaires?.replace(/"/gu, '').replace(/\n/gu, ''),
-  lieuMediationNumerique.presentation_resume?.replace(/"/gu, '＂').replace(/\n/gu, ''),
-  lieuMediationNumerique.presentation_detail?.replace(/"/gu, '＂').replace(/\n/gu, ''),
+  lieuMediationNumerique.horaires?.replace(/"/g, '').replace(/\n/g, ''),
+  lieuMediationNumerique.presentation_resume?.replace(/"/g, '＂').replace(/\n/g, ''),
+  lieuMediationNumerique.presentation_detail?.replace(/"/g, '＂').replace(/\n/g, ''),
   lieuMediationNumerique.source,
   lieuMediationNumerique.itinerance,
   lieuMediationNumerique.structure_parente,
   lieuMediationNumerique.date_maj,
-  lieuMediationNumerique.services?.replace(/"/gu, '＂').replace(/\n/gu, ''),
+  lieuMediationNumerique.services?.replace(/"/g, '＂').replace(/\n/g, ''),
   lieuMediationNumerique.publics_specifiquement_adresses,
   lieuMediationNumerique.prise_en_charge_specifique,
   lieuMediationNumerique.frais_a_charge,
   lieuMediationNumerique.dispositif_programmes_nationaux,
   lieuMediationNumerique.formations_labels,
-  lieuMediationNumerique.autres_formations_labels?.replace(/"/gu, '＂').replace(/\n/gu, ''),
+  lieuMediationNumerique.autres_formations_labels?.replace(/"/g, '＂').replace(/\n/g, ''),
   lieuMediationNumerique.modalites_acces,
   lieuMediationNumerique.modalites_accompagnement,
   lieuMediationNumerique.fiche_acces_libre,
