@@ -1,4 +1,3 @@
-/* eslint-disable-next-line @typescript-eslint/no-restricted-imports */
 import * as fs from 'fs';
 
 export const throwWriteFileError = (writeFileError: unknown): void => {
@@ -10,7 +9,7 @@ export const throwWriteFileError = (writeFileError: unknown): void => {
 export const noEmptyCell = <T>(_: string, cell: T): T | undefined => (cell === '' ? undefined : cell);
 
 export const createFolderIfNotExist = (folderPath: string): string => {
-  !fs.existsSync(folderPath) && fs.mkdirSync(folderPath, { recursive: true });
+  if (!fs.existsSync(folderPath)) fs.mkdirSync(folderPath, { recursive: true });
 
   return folderPath;
 };
