@@ -20,14 +20,14 @@ const DUPLICATION_COMPARISON_HEADINGS: string = [
 const duplicationComparisonLineFor = (duplicationComparison: DuplicationComparison): string =>
   [
     duplicationComparison.score,
-    duplicationComparison.typologie1?.replace(';', ','),
-    duplicationComparison.typologie2?.replace(';', ','),
+    duplicationComparison.typologie1?.replace(/\|/g, ','),
+    duplicationComparison.typologie2?.replace(/\|/g, ','),
     duplicationComparison.nomScore,
-    duplicationComparison.nom1.replace(';', ''),
-    duplicationComparison.nom2.replace(';', ''),
+    duplicationComparison.nom1.replace(/;/g, ''),
+    duplicationComparison.nom2.replace(/;/g, ''),
     duplicationComparison.adresseScore,
-    duplicationComparison.adresse1.replace(';', ''),
-    duplicationComparison.adresse2.replace(';', ''),
+    duplicationComparison.adresse1.replace(/;/g, ''),
+    duplicationComparison.adresse2.replace(/;/g, ''),
     duplicationComparison.distanceScore,
     duplicationComparison.localisation1,
     duplicationComparison.localisation2,
