@@ -1,6 +1,11 @@
 import { SchemaLieuMediationNumerique } from '@gouvfr-anct/lieux-de-mediation-numerique';
-import { Groups, MergedLieuxByGroupMap } from '../steps';
+import { DuplicationComparison, Groups, MergedLieuxByGroupMap } from '../steps';
 
 export type DeduplicationRepository = {
-  save: (groups: Groups, merged: MergedLieuxByGroupMap, lieuxToDeduplicate?: SchemaLieuMediationNumerique[]) => Promise<void>;
+  save: (
+    groups: Groups,
+    merged: MergedLieuxByGroupMap,
+    lieuxToDeduplicate?: SchemaLieuMediationNumerique[],
+    duplications?: DuplicationComparison[]
+  ) => Promise<void>;
 };
