@@ -1,11 +1,11 @@
-import { SchemaLieuMediationNumerique, Typologie } from '@gouvfr-anct/lieux-de-mediation-numerique';
+import { SchemaLieuMediationNumerique, Typologie, Typologies } from '@gouvfr-anct/lieux-de-mediation-numerique';
 import { ratio } from 'fuzzball';
 
 export type Duplicate = { id: string; distanceScore: number; nomFuzzyScore: number; voieFuzzyScore: number };
 
 export type LieuDuplications = { id: string; duplicates: Duplicate[] };
 
-export type CommuneDuplications = { codePostal: string; lieux: LieuDuplications[] };
+export type CommuneDuplications = { code_insee: string; lieux: LieuDuplications[] };
 
 const sameSource = (lieu: SchemaLieuMediationNumerique, lieuToDeduplicate: SchemaLieuMediationNumerique): boolean =>
   lieu.source === lieuToDeduplicate.source;
@@ -13,38 +13,41 @@ const sameSource = (lieu: SchemaLieuMediationNumerique, lieuToDeduplicate: Schem
 const sameId = (lieu: SchemaLieuMediationNumerique, lieuToDeduplicate: SchemaLieuMediationNumerique): boolean =>
   lieu.id === lieuToDeduplicate.id;
 
-const sameCodePostal = (lieu: SchemaLieuMediationNumerique, lieuToDeduplicate: SchemaLieuMediationNumerique): boolean =>
-  lieu.code_postal === lieuToDeduplicate.code_postal;
+const sameCodeInsee = (lieu: SchemaLieuMediationNumerique, lieuToDeduplicate: SchemaLieuMediationNumerique): boolean =>
+  lieu.code_insee === lieuToDeduplicate.code_insee;
 
-const hasRFSCompatibleTypology = (lieu: SchemaLieuMediationNumerique): boolean =>
-  [`${Typologie.RFS}`, `${Typologie.PIMMS}`].includes(lieu.typologie ?? 'NO_TYPOLOGY');
+const compatibleTypologies: [Typologie, Typologie][] = [[Typologie.RFS, Typologie.PIMMS]];
 
-const isCompatibleForFranceServices = (
-  lieu: SchemaLieuMediationNumerique,
-  lieuToDeduplicate: SchemaLieuMediationNumerique
-): boolean =>
-  (hasRFSCompatibleTypology(lieu) && hasRFSCompatibleTypology(lieuToDeduplicate)) ||
-  (/france services?/giu.test(lieu.nom.toLowerCase()) && /france services?/giu.test(lieuToDeduplicate.nom.toLowerCase()));
+const toTypologies = ({ typologie }: { typologie?: string }): Typologies =>
+  Typologies((typologie?.split('|') as Typologies) ?? []);
 
-const compatibilitySpecialCases = (
-  lieu: SchemaLieuMediationNumerique,
-  lieuToDeduplicate: SchemaLieuMediationNumerique
-): boolean =>
-  isCompatibleForFranceServices(lieu, lieuToDeduplicate)
-    ? true
-    : lieuToDeduplicate.typologie !== Typologie.RFS && lieu.typologie !== Typologie.RFS;
+const allEmpty = (lieuTypologies: Typologies, lieuToDeduplicateTypologies: Typologies) =>
+  lieuTypologies.length === 0 && lieuToDeduplicateTypologies.length === 0;
+
+const hasSameTypologieIn = (lieuToDeduplicateTypologies: Typologies) => (typologie: Typologie) =>
+  lieuToDeduplicateTypologies.includes(typologie);
+
+const hasCompatibleTypologiesFor =
+  (lieuTypologies: Typologies, lieuToDeduplicateTypologies: Typologies) =>
+  ([typologyA, typologyB]: [Typologie, Typologie]) =>
+    (lieuTypologies.includes(typologyA) && lieuToDeduplicateTypologies.includes(typologyB)) ||
+    (lieuTypologies.includes(typologyB) && lieuToDeduplicateTypologies.includes(typologyA));
+
+const sameTypologie = (lieuTypologies: Typologies, lieuToDeduplicateTypologies: Typologies): boolean =>
+  allEmpty(lieuTypologies, lieuToDeduplicateTypologies) ||
+  lieuTypologies.some(hasSameTypologieIn(lieuToDeduplicateTypologies)) ||
+  compatibleTypologies.some(hasCompatibleTypologiesFor(lieuTypologies, lieuToDeduplicateTypologies));
 
 const onlyPotentialDuplicates =
   (lieuToDeduplicate: SchemaLieuMediationNumerique, allowInternalMerge: boolean) =>
   (lieu: SchemaLieuMediationNumerique): boolean =>
-    sameCodePostal(lieu, lieuToDeduplicate) &&
+    sameCodeInsee(lieu, lieuToDeduplicate) &&
     !sameId(lieu, lieuToDeduplicate) &&
     (allowInternalMerge || !sameSource(lieu, lieuToDeduplicate)) &&
-    compatibilitySpecialCases(lieu, lieuToDeduplicate);
+    sameTypologie(toTypologies(lieu), toTypologies(lieuToDeduplicate));
 
 const MINIMAL_CARTESIAN_DISTANCE: 0.0004 = 0.0004 as const;
 
-/* eslint-disable-next-line no-mixed-operators */
 const computeDistanceScore = (cartesianDistance: number): number => 100 / (cartesianDistance * 1000 + 1);
 
 const distanceScore = (cartesianDistance: number): number =>
@@ -55,7 +58,6 @@ const hasDefinedCoordinates = (
 ): lieu is SchemaLieuMediationNumerique & { latitude: number; longitude: number } =>
   lieu.latitude != null && lieu.longitude != null;
 
-/* eslint-disable-next-line no-mixed-operators */
 const pythagore = (x1: number, x2: number, y1: number, y2: number): number => Math.sqrt((x1 - x2) ** 2 + (y1 - y2) ** 2);
 
 const cartesianDistanceBetween = (lieu: SchemaLieuMediationNumerique, cible: SchemaLieuMediationNumerique): number =>
@@ -87,36 +89,42 @@ const appendCommuneDuplications =
     lieuToDeduplicate: SchemaLieuMediationNumerique,
     duplications: CommuneDuplications[],
     allowInternalMerge: boolean
-  ): CommuneDuplications[] => [
-    ...duplications,
-    { codePostal: lieuToDeduplicate.code_postal, lieux: [toLieuDuplications(lieuToDeduplicate, lieux, allowInternalMerge)] }
-  ];
+  ): CommuneDuplications[] =>
+    lieuToDeduplicate.code_insee != null
+      ? [
+          ...duplications,
+          {
+            code_insee: lieuToDeduplicate.code_insee,
+            lieux: [toLieuDuplications(lieuToDeduplicate, lieux, allowInternalMerge)]
+          }
+        ]
+      : duplications;
 
 const toUpdatedCommuneDuplications =
   (lieux: SchemaLieuMediationNumerique[]) =>
-  (lieu: SchemaLieuMediationNumerique, duplicationsWithSameCodePostal: CommuneDuplications, allowInternalMerge: boolean) =>
+  (lieu: SchemaLieuMediationNumerique, duplicationsWithSameCodeInsee: CommuneDuplications, allowInternalMerge: boolean) =>
   (communeDuplications: CommuneDuplications): CommuneDuplications =>
-    communeDuplications.codePostal === lieu.code_postal
+    communeDuplications.code_insee === lieu.code_insee
       ? {
-          codePostal: lieu.code_postal,
-          lieux: [...duplicationsWithSameCodePostal.lieux, toLieuDuplications(lieu, lieux, allowInternalMerge)]
+          code_insee: lieu.code_insee,
+          lieux: [...duplicationsWithSameCodeInsee.lieux, toLieuDuplications(lieu, lieux, allowInternalMerge)]
         }
       : communeDuplications;
 
-const withSameCodePostal =
+const withSameCodeInsee =
   (lieu: SchemaLieuMediationNumerique) =>
-  ({ codePostal }: CommuneDuplications): boolean =>
-    codePostal === lieu.code_postal;
+  ({ code_insee }: CommuneDuplications): boolean =>
+    code_insee === lieu.code_insee;
 
 const toCommunesDuplications =
   (lieux: SchemaLieuMediationNumerique[], allowInternalMerge: boolean) =>
   (duplications: CommuneDuplications[], lieuToDeduplicate: SchemaLieuMediationNumerique): CommuneDuplications[] =>
-    ((duplicationsWithSameCodePostal?: CommuneDuplications): CommuneDuplications[] =>
-      duplicationsWithSameCodePostal == null
+    ((duplicationsWithSameCodeInsee?: CommuneDuplications): CommuneDuplications[] =>
+      duplicationsWithSameCodeInsee == null
         ? appendCommuneDuplications(lieux)(lieuToDeduplicate, duplications, allowInternalMerge)
         : duplications.map(
-            toUpdatedCommuneDuplications(lieux)(lieuToDeduplicate, duplicationsWithSameCodePostal, allowInternalMerge)
-          ))(duplications.find(withSameCodePostal(lieuToDeduplicate)));
+            toUpdatedCommuneDuplications(lieux)(lieuToDeduplicate, duplicationsWithSameCodeInsee, allowInternalMerge)
+          ))(duplications.find(withSameCodeInsee(lieuToDeduplicate)));
 
 const onlyWithDuplicates = ({ duplicates }: LieuDuplications): boolean => duplicates.length > 0;
 
@@ -132,8 +140,8 @@ const toDuplicatesWithout =
 const invalidDuplicatesIds = ({ lieux }: CommuneDuplications): string[] =>
   lieux.filter(onlyWithoutDuplicates).map((lieu: LieuDuplications): string => lieu.id);
 
-const removeLieuxFrom = ({ codePostal, lieux }: CommuneDuplications, ids: string[]): CommuneDuplications => ({
-  codePostal,
+const removeLieuxFrom = ({ code_insee, lieux }: CommuneDuplications, ids: string[]): CommuneDuplications => ({
+  code_insee,
   lieux: lieux.map(toDuplicatesWithout(ids)).filter(onlyWithDuplicates)
 });
 
