@@ -1,5 +1,4 @@
-/* eslint-disable @typescript-eslint/naming-convention, camelcase */
-
+import { describe, it, expect } from 'vitest';
 import {
   SchemaLieuMediationNumerique,
   Service,
@@ -58,6 +57,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -71,6 +71,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -95,6 +96,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -115,6 +117,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -128,6 +131,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -153,6 +157,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -174,6 +179,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -188,6 +194,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -212,6 +219,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -233,6 +241,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -247,6 +256,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -271,6 +281,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -292,6 +303,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -307,6 +319,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -332,6 +345,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -354,6 +368,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -369,6 +384,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -394,6 +410,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -416,6 +433,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -432,6 +450,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -458,6 +477,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -481,6 +501,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -496,6 +517,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -521,6 +543,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -543,6 +566,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -558,6 +582,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -584,6 +609,7 @@ describe('remove duplicates', (): void => {
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
             commune: 'Durtal',
+            code_insee: '49127',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
             date_maj: '2023-05-03',
@@ -605,6 +631,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -620,6 +647,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -645,6 +673,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -671,6 +700,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -686,6 +716,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -711,6 +742,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -733,6 +765,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -747,6 +780,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -771,6 +805,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'DURTAL',
             latitude: 47.671271,
             longitude: -0.256457,
@@ -792,6 +827,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -805,6 +841,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -829,6 +866,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -850,6 +888,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -863,6 +902,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -888,6 +928,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -908,6 +949,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -922,6 +964,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -946,6 +989,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'DURTAL',
             latitude: 47.671271,
             longitude: -0.256457,
@@ -966,6 +1010,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -979,6 +1024,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -1003,6 +1049,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
@@ -1023,6 +1070,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -1036,6 +1084,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -1060,6 +1109,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'DURTAL',
             latitude: 47.671271,
             longitude: -0.256457,
@@ -1080,6 +1130,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'Durtal',
         latitude: 47.6699154795,
         longitude: -0.2551539846,
@@ -1094,6 +1145,7 @@ describe('remove duplicates', (): void => {
         nom: 'France Services Durtal',
         adresse: '11 rue Joseph Cugnot',
         code_postal: '49430',
+        code_insee: '49127',
         commune: 'DURTAL',
         latitude: 47.671271,
         longitude: -0.256457,
@@ -1119,6 +1171,7 @@ describe('remove duplicates', (): void => {
             nom: 'France Services Durtal',
             adresse: '11 rue Joseph Cugnot',
             code_postal: '49430',
+            code_insee: '49127',
             commune: 'Durtal',
             latitude: 47.6699154795,
             longitude: -0.2551539846,
