@@ -1,5 +1,6 @@
+import { describe, it, expect } from 'vitest';
 import { Localisation } from '@gouvfr-anct/lieux-de-mediation-numerique';
-import { Polygon } from '@turf/turf';
+import { type Polygon } from 'geojson';
 import { QpvShapesMap } from '../../fields';
 import { isInQpv } from './is-in-qpv';
 
