@@ -1,5 +1,4 @@
-/* eslint-disable @typescript-eslint/naming-convention, camelcase */
-
+import { describe, it, expect } from 'vitest';
 import { SourceMap, sourceMapFromTransfer, SourceTransfer } from './source.transfer';
 
 describe('source transfer', (): void => {
