@@ -1,5 +1,4 @@
-/* eslint-disable @typescript-eslint/naming-convention, camelcase */
-
+import { describe, it, expect } from 'vitest';
 import { zrrMapFromTransfer, ZrrTransfer } from './zrr.transfer';
 
 describe('zrr transfer', (): void => {
