@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/naming-convention, camelcase */
-
 import { Adresse } from '@gouvfr-anct/lieux-de-mediation-numerique';
 import { DataSource, LieuxMediationNumeriqueMatching } from '../../input';
 import { getNewCommune } from './anciennes-communes';
@@ -24,7 +22,7 @@ type SourceAddress = {
 const nouvelleCaledonieException = (codePostal: string): boolean => codePostal.startsWith('98');
 
 export const complementAdresseIfAny = (complementAdresse?: string): { complement_adresse?: string } =>
-  complementAdresse == null ? {} : { complement_adresse: complementAdresse.replace(/\s+/gu, ' ').trim() };
+  complementAdresse == null ? {} : { complement_adresse: complementAdresse.replace(/\s+/g, ' ').trim() };
 
 const codeInseeIfAny = (code_insee?: string): { code_insee?: string } => (code_insee == null ? {} : { code_insee });
 
