@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/naming-convention, camelcase */
-
 import { Contact, Courriel, Url } from '@gouvfr-anct/lieux-de-mediation-numerique';
 import { LieuxMediationNumeriqueMatching, DataSource } from '../../input';
 import { Recorder } from '../../report';
@@ -7,7 +5,7 @@ import { cleanOperations, CleanOperation } from './clean-operations';
 
 type FixedContact = DataSource | undefined;
 
-const toInternationalFormat = (phone: string): string => (/^0\d{9}$/u.test(phone) ? `+33${phone.slice(1)}` : phone);
+const toInternationalFormat = (phone: string): string => (/^0\d{9}$/.test(phone) ? `+33${phone.slice(1)}` : phone);
 
 const telephoneField = (telephone?: number | string): Pick<Contact, 'telephone'> =>
   telephone == null
@@ -16,7 +14,7 @@ const telephoneField = (telephone?: number | string): Pick<Contact, 'telephone'>
         telephone: toInternationalFormat(
           telephone
             .toString()
-            .replace(/[\s,.-]/gu, '')
+            .replace(/[\s,.-]/g, '')
             .replace('(0)', '')
             .trim()
         )
@@ -44,6 +42,7 @@ const shouldApplyFix = (cleanOperation: CleanOperation, property?: string): bool
 const applyRemoveFix =
   (recorder: Recorder) =>
   (cleanOperation: CleanOperation, valueToFix: string, source: DataSource): DataSource => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { [cleanOperation.field]: removedProperty, ...filteredProperties }: DataSource = source;
     recorder.fix({
       apply: cleanOperation.name,
