@@ -1,5 +1,4 @@
-/* eslint-disable @typescript-eslint/no-unnecessary-condition  */
-import { OsmDaysOfWeek, OsmOpeningHours, toOsmOpeningHours } from '@gouvfr-anct/timetable-to-osm-opening-hours';
+import { OsmDaysOfWeek, OsmOpeningHours, fromTimetableOsmOpeningHours } from '@gouvfr-anct/timetable-to-osm-opening-hours';
 import { LieuxMediationNumeriqueMatching, DataSource } from '../../input';
 import { toOsmHours } from '../../to-osm-hours/to-osm-hours';
 import { InvalidHoursError } from './errors/invalid-hours-error';
@@ -9,10 +8,10 @@ import { openingHoursFromWeek } from './opening-hours-from-week';
 const SEMAINE_IMPAIRE: string = 'week 1-53/2 ';
 const SEMAINE_PAIRE: string = 'week 2-52/2 ';
 
-const OPENING_HOURS_REGEXP: RegExp = /^\d{2}:\d{2}-\d{2}:\d{2}(?:,\d{2}:\d{2}-\d{2}:\d{2})?$/u;
+const OPENING_HOURS_REGEXP: RegExp = /^\d{2}:\d{2}-\d{2}:\d{2}(?:,\d{2}:\d{2}-\d{2}:\d{2})?$/;
 
 const OSM_OPENING_HOURS_TRIVIAL_REGEXP: RegExp =
-  /^(?:(?:Mo|Tu|We|Th|Fr|Sa|Su)(?:[-,](?:Mo|Tu|We|Th|Fr|Sa|Su))?\s)?(?:[0-1]\d|2[0-3]):[0-5]\d-(?:[0-1]\d|2[0-3]):[0-5]\d.*/u;
+  /^(?:(?:Mo|Tu|We|Th|Fr|Sa|Su)(?:[-,](?:Mo|Tu|We|Th|Fr|Sa|Su))?\s)?(?:[0-1]\d|2[0-3]):[0-5]\d-(?:[0-1]\d|2[0-3]):[0-5]\d.*/;
 
 const throwInvalidHours = (osmHours: string, day: OsmDaysOfWeek, hours: string): OsmOpeningHours => {
   throw new InvalidHoursError(osmHours, hours, day);
@@ -32,7 +31,7 @@ const processDay = (day: OsmDaysOfWeek, hours?: string): [] | [OsmOpeningHours] 
 
 const openingHoursFromDays = (matching: LieuxMediationNumeriqueMatching, source: DataSource): OsmOpeningHoursString =>
   osmOpeningHoursString(
-    toOsmOpeningHours([
+    fromTimetableOsmOpeningHours([
       ...(matching.horaires?.jours?.reduce(
         (
           processedDay: OsmOpeningHours[],
