@@ -1,6 +1,4 @@
-/* eslint-disable @typescript-eslint/naming-convention, camelcase */
-
-/* eslint-disable-next-line @typescript-eslint/no-shadow */
+import { describe, it, expect } from 'vitest';
 import { Presentation } from '@gouvfr-anct/lieux-de-mediation-numerique';
 import { LieuxMediationNumeriqueMatching, DataSource } from '../../input';
 import { processPresentation } from './presentation.field';
